new test for new loc property observers
